Improve ToneGeneratorNode (#242)

+++ deps/include/stk/DelayA.h
@@ -44,7 +44,7 @@ public:
   void clear( void );
 
   //! Get the maximum delay-line length.
-  unsigned long getMaximumDelay( void ) { return inputs_.size() - 1; };
+  size_t getMaximumDelay( void ) { return inputs_.size() - 1; };
   
   //! Set the maximum delay-line length.
   /*!
